[REFACTOR] SignIn/Up 뷰모델 분리 및 fcmToken 초기화 구현 (#667)
