+++ app/src/main/AndroidManifest.xml
@@ -143,7 +143,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".ui.sign.SignActivity"
+            android:name=".ui.sign.SignUpActivity"
             android:configChanges="orientation"
             android:exported="true"
             android:screenOrientation="portrait" />
