--- app/src/main/java/org/sopt/havit/ui/sign/SignViewModel.kt
@@ -1,210 +1,0 @@
-package org.sopt.havit.ui.sign
-
-import android.util.Log
-import android.view.View
-import androidx.lifecycle.*
-import com.kakao.sdk.auth.model.OAuthToken
-import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.launch
-import org.sopt.havit.R
-import org.sopt.havit.data.remote.SignInResponse
-import org.sopt.havit.data.remote.SignUpRequest
-import org.sopt.havit.domain.entity.NetworkState
-import org.sopt.havit.domain.repository.AuthRepository
-import org.sopt.havit.util.Event
-import retrofit2.HttpException
-import javax.inject.Inject
-
-@HiltViewModel
-class SignViewModel @Inject constructor(
-    private val authRepository: AuthRepository
-) : ViewModel() {
-
-    companion object {
-        const val SPLASH_FROM_SHARE = true
-        const val SPLASH_NORMAL_FLOW = false
-    }
-
-    var isServerNetwork = MutableLiveData<NetworkState>()
-
-    var loginGuidVisibility = MutableLiveData(false)
-        private set
-
-    private var _isMoveToNextOrBack = MutableLiveData<Event<Boolean>>()
-    var isMoveToNextOrBack: LiveData<Event<Boolean>> = _isMoveToNextOrBack
-
-
-    private var _isAlreadyUser = MutableLiveData<Event<SignInResponse>>()
-    var isAlreadyUser: LiveData<Event<SignInResponse>> = _isAlreadyUser
-
-    private var _isNeedScopes = MutableLiveData<Event<Boolean>>()
-    var isNeedScopes: LiveData<Event<Boolean>> = _isNeedScopes
-
-    private var _isReadyUser = MutableLiveData<Event<Boolean>>()
-    var isReadyUser: LiveData<Event<Boolean>> = _isReadyUser
-
-    private var _accessToken = MutableLiveData<String>()
-    var accessToken: LiveData<String> = _accessToken
-
-    private val _fcmToken = MutableLiveData<String>()
-
-    private val _kakaoToken = MutableLiveData<String>()
-
-    var nickName = MutableLiveData<String>()
-
-    var _havitUser = MutableLiveData<SignUpRequest>()
-
-    var isTosUseCheck = MutableLiveData(false)
-    var isTosInfoCheck = MutableLiveData(false)
-    var isTosEventCheck = MutableLiveData(false)
-
-    var isAllCheck = MediatorLiveData<Boolean>().apply {
-        addSource(isTosUseCheck) { value = checkTosAll() }
-        addSource(isTosInfoCheck) { value = checkTosAll() }
-        addSource(isTosEventCheck) { value = checkTosAll() }
-    }
-
-    private fun checkTosAll(): Boolean {
-        return isTosUseCheck.value == true && isTosInfoCheck.value == true && isTosEventCheck.value == true
-    }
-
-    var isNextClick = MediatorLiveData<Boolean>().apply {
-        addSource(isAllCheck) {
-            value =
-                isAllCheck.value == true || (isTosUseCheck.value == true && isTosInfoCheck.value == true)
-        }
-    }
-
-
-    val kakaoLoginCallback: (OAuthToken?, Throwable?) -> Unit = { token, error ->
-        if (error != null) {
-            Log.d("TAG", "카카오계정으로 로그인 실패", error)
-        } else if (token != null) {
-            Log.d("TAG", "카카오계정으로 로그인 성공 ${token.accessToken}")
-            setKakaoToken(token.accessToken)
-            setFcmToken()
-            setNeedScopes()
-        }
-    }
-
-    fun setLoginGuideVisibility(boolean: Boolean) {
-        loginGuidVisibility.value = boolean
-    }
-
-    fun setNeedScopes() {
-        _isNeedScopes.value = Event(true)
-    }
-
-    fun setReadyUser() {
-        _isReadyUser.value = Event(true)
-    }
-
-    fun setMoveToNextOrBack(move: Boolean) {
-        _isMoveToNextOrBack.value = Event(move)
-    }
-
-    fun setKakaoUserInfo(age: Int, email: String, gender: String, nickName: String) {
-        val user = SignUpRequest(
-            age,
-            email,
-            requireNotNull(_fcmToken.value),
-            gender,
-            requireNotNull(_kakaoToken.value),
-            nickName
-        )
-        _havitUser.value = user
-        setNickName(nickName)
-    }
-
-    fun postSignUp() {
-        viewModelScope.launch {
-            kotlin.runCatching {
-                authRepository.postSignUp(
-                    requireNotNull(_havitUser.value?.email),
-                    requireNotNull(nickName.value),
-                    requireNotNull(_havitUser.value?.age),
-                    requireNotNull(_havitUser.value?.gender),
-                    requireNotNull(_fcmToken.value),
-                    requireNotNull(_kakaoToken.value)
-                )
-            }.onSuccess {
-                if (it.success) {
-                    authRepository.saveAccessToken(requireNotNull(it.data.accessToken))
-                    _accessToken.postValue(it.data.accessToken)
-                    isServerNetwork.postValue(NetworkState.SUCCESS)
-                }
-            }.onFailure {
-                isServerNetwork.postValue(NetworkState.FAIL)
-                Log.d("error", "${(it as? HttpException)?.message}")
-            }
-        }
-    }
-
-    fun getSignIn() {
-        viewModelScope.launch {
-            kotlin.runCatching {
-                authRepository.getSignIn(
-                    requireNotNull(_fcmToken.value), requireNotNull(_kakaoToken.value)
-                )
-            }.onSuccess {
-                if (it.success) {
-                    _isAlreadyUser.postValue(Event(it))
-                    if (it.data.isAlreadyUser == null) {
-                        authRepository.saveAccessToken(
-                            requireNotNull(
-                                it.data.accessToken
-                            )
-                        )
-                    }
-                }
-            }.onFailure { Log.d("error", "${(it as? HttpException)?.message}") }
-        }
-    }
-
-    fun setFcmToken() {
-        authRepository.getFcmToken { token -> _fcmToken.value = token }
-    }
-
-    fun setKakaoToken(token: String) {
-        _kakaoToken.value = token
-        authRepository.saveKakaoToken(token)
-    }
-
-    fun getKakaoToken() {
-        _kakaoToken.value = authRepository.getKakaoToken()
-    }
-
-    fun clickTosListener(view: View) {
-        when (view.id) {
-            R.id.iv_tos_info -> isTosInfoCheck.value = !isTosInfoCheck.value!!
-            R.id.iv_tos_use -> isTosUseCheck.value = !isTosUseCheck.value!!
-            R.id.iv_tos_event -> isTosEventCheck.value = !isTosEventCheck.value!!
-            R.id.iv_tos_all, R.id.tv_tos_all -> {
-                setAllTosClick()
-            }
-        }
-
-    }
-
-    private fun setAllTosClick() {
-        if (isAllCheck.value == false) {
-            isAllCheck.value = true
-            isTosUseCheck.value = true
-            isTosInfoCheck.value = true
-            isTosEventCheck.value = true
-        } else {
-            isAllCheck.value = false
-            isTosUseCheck.value = false
-            isTosInfoCheck.value = false
-            isTosEventCheck.value = false
-        }
-    }
-
-    fun setNickName(name: String) {
-        nickName.value = name
-    }
-
-    fun setClearNickName() {
-        nickName.value = ""
-    }
-}
