+++ app/src/main/res/layout/activity_sign_up.xml
@@ -6,7 +6,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.sign.SignActivity">
+        tools:context=".ui.sign.SignUpActivity">
 
         <androidx.viewpager2.widget.ViewPager2
             android:id="@+id/vp_sign"
