+++ app/src/main/res/layout/activity_splash_with_sign.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="main"
-            type="org.sopt.havit.ui.sign.SignViewModel" />
+            type="org.sopt.havit.ui.sign.SignInViewModel" />
 
         <import type="android.view.View" />
