+++ app/src/main/res/layout/fragment_add_nick_name.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="vm"
-            type="org.sopt.havit.ui.sign.SignViewModel" />
+            type="org.sopt.havit.ui.sign.SignUpViewModel" />
 
     </data>
