+++ app/src/main/res/layout/fragment_add_tos.xml
@@ -13,7 +13,7 @@
 
         <variable
             name="vm"
-            type="org.sopt.havit.ui.sign.SignViewModel" />
+            type="org.sopt.havit.ui.sign.SignUpViewModel" />
 
     </data>
