+++ build.gradle
@@ -6,10 +6,10 @@ buildscript {
     }
     dependencies {
         classpath "com.android.tools.build:gradle:7.0.4"
-        classpath 'org.jetbrains.kotlin:kotlin-gradle-plugin:1.6.21'
+        classpath 'org.jetbrains.kotlin:kotlin-gradle-plugin:1.7.0'
         classpath "androidx.navigation:navigation-safe-args-gradle-plugin:2.4.2"
         classpath 'com.google.gms:google-services:4.3.10'
-        classpath 'com.google.dagger:hilt-android-gradle-plugin:2.39.1'
+        classpath 'com.google.dagger:hilt-android-gradle-plugin:2.42'
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
     }
